refactor(core): Extract file filtering constants from Config
Move FileFilteringOptions and related default constants from config.ts to a new constants.ts file. This decouples these shared data structures from the Config class, preventing potential circular dependencies when tools (like GlobTool) need these constants and are themselves imported by agents that the Config class depends on.
